Added the new semi-dictionary type and worked it into the existing hierarchy
Also removed some unnecessary interface-specific re-implementations of the indexer

The normal indexer works fine to implement those interface functions
