+++ CsharpExtras/Map/Dictionary/Variant/Semi/IInKeyOutValueSemiDictionary.cs
@@ -13,22 +13,4 @@ namespace CsharpExtras.Map.Dictionary.Variant.Semi
         ISuccessTuple<V> TryGetValue(K key);
     }
 
-    public interface ISuccessTuple<out V>
-    {
-        bool WasSuccessful { get; }
-        V Value { get; }
-    }
-
-    public class SuccessTupleImpl<V> : ISuccessTuple<V>
-    {
-        public bool WasSuccessful { get; }
-        public V Value {get;}
-
-        public SuccessTupleImpl(bool wasSuccessful, V value)
-        {
-            WasSuccessful = wasSuccessful;
-            Value = value;
-        }
-    }
-
 }
