+++ CsharpExtras/Map/Dictionary/Variant/Semi/ISuccessTuple.cs
@@ -1,0 +1,9 @@
+﻿namespace CsharpExtras.Map.Dictionary.Variant.Semi
+{
+    public interface ISuccessTuple<out V>
+    {
+        bool WasSuccessful { get; }
+        V Value { get; }
+    }
+
+}
