+++ CsharpExtras/Map/Dictionary/Variant/Semi/SuccessTupleImpl.cs
@@ -1,0 +1,15 @@
+﻿namespace CsharpExtras.Map.Dictionary.Variant.Semi
+{
+    public class SuccessTupleImpl<V> : ISuccessTuple<V>
+    {
+        public bool WasSuccessful { get; }
+        public V Value {get;}
+
+        public SuccessTupleImpl(bool wasSuccessful, V value)
+        {
+            WasSuccessful = wasSuccessful;
+            Value = value;
+        }
+    }
+
+}
